feat: manager system structure
