+++ main.py
@@ -1,42 +1,25 @@
 import os
 import discord
 import asyncio
-from discord.ext import commands
+
 from dotenv import load_dotenv
-
-
-class Client(commands.Bot):
-    async def on_ready(self):
-        print(f"Logged on as {self.user}!")
-
-    async def on_message(self, message):
-        # Ensures commands work properly, wont process any commands without it.
-        await self.process_commands(message)
-
-        print(f"Message from {message.author}: {message.content}")
+from src.managers.cog_system import CogManager
+from src.managers.discord_client import DiscordClient
 
 
 # New python API update requires these
 intents = discord.Intents.default()
 intents.message_content = True
 
-client = Client(command_prefix="!", intents=intents, help_command=None)
-
-
-async def load_cogs():
-    # TODO: Ensure this is only loaded in debug mode, not in live
-    # Remote code execution module for debugging purposes
-    await client.load_extension("jishaku")
-
-    for filename in os.listdir("./cogs"):
-        if filename.endswith(".py"):
-            await client.load_extension(f"cogs.{filename[:-3]}")
+client = DiscordClient(command_prefix="!", intents=intents, help_command=None)
 
 
 async def main():
-    await load_cogs()
+    cog_system = CogManager(client)
+    await cog_system.reg_cogs()
 
     load_dotenv()
+    # TODO: Check if file is present first
     token = os.getenv("TOKEN")
     if token is None:
         print("No token was provided in .env, shutting down...")
