+++ src/managers/cog_system.py
@@ -1,0 +1,18 @@
+import os
+
+
+class CogManager:
+    def __init__(self, client):
+        self.client = client
+
+    async def reg_debug_cogs(self):
+        # Remote code execution module for debugging purposes
+        await self.client.load_extension("jishaku")
+
+    async def reg_cogs(self):
+        # TODO: Ensure this is only loaded in debug mode, not in live
+        self.reg_debug_cogs()
+
+        for filename in os.listdir("./src/cogs"):
+            if filename.endswith(".py"):
+                await self.client.load_extension(f"src.cogs.{filename[:-3]}")
